feat: new rule attribute-names
This introduces a new rule: `attribute-names`.

The rule enforces that all attributes are lowercase if the associated
property is not.

For example:

```ts
class Foo extends LitElement {
  @property()
  camelCase = 'foo';
}
```

This will fail since the default attribute will be `camelCase`, and
therefore not lowercase.

This should instead be:

```ts
@property({attribute: 'camel-case'})
```

Though in this rule, for now, the fact it is snake-case is a preference
rather than being enforced by the rule.