test: simplify config assertions
A quick change to dumb the config tests down so they basically assert
against the exported structure rather than an integration-like test.

May still be worth an integration test in future.

Previously, the 2nd test case was actually passing due to an unrelated
error (parse error, too low of an ecma version), which was satisfying
the later assertion.